requirements.txt
python-docx
customtkinter
Pillow
pyinstaller
tkinterdnd2
pywin32